Implement general mobility and material scoring
The new constant at the top of the file `PieceVectorPositions` tells us
the position that each piece should take in a vector. I suppose this
doesn't need to be enforced everywhere, but if you are implementing a
feature that has some value for all pieces, it will be useful.

I also added docstrings/type hints to any function I touched. While not
strictly necessary, they're nice to have (and make our project look a
bit nicer, too).